feat: add invalid input message

+++ deleteFile.js
@@ -2,9 +2,15 @@ import fs from 'fs/promises';
 import { absolutePath } from './cdDir.js';
 
 export const deleteFile = (dir, file) => {
-  const filePath = absolutePath(dir, file);
+  if (!file) {
+    console.log('Invalid input');
+  }
+  else {
+    const filePath = absolutePath(dir, file);
     
-  fs.rm(filePath)
-    .then(() => console.log('Success'))
-    .catch((err) => console.log('FS operation failed'));
+    fs.rm(filePath)
+      .then(() => console.log('Success'))
+      .catch((err) => console.log('FS operation failed'));
+  }
+
 };
